Many changes (add tray icon, hide-on-minimize, other options)
- Change project name from 'brightness' to 'LinkBrightness'.

- Change assembly name from 'Link Brightness' to 'LinkBrightness'.

- Add website to assembly trademark/company.

- Add myself to assembly copyright.

- Bump assembly version number to 1.1.*.

- Add a program icon.

The GIMP file used to create it is LinkBrightness.xcf.

- Add a system tray icon using the program icon.

- Add the ability to show/hide the console via tray icon double-click.

The tray icon tooltip: "double-click to show/hide".

- Add a context menu to the tray icon (items: About, Exit).

- Add an AboutBox that explains what the program does.

- New option /hide_on_minimize.

This option starts an event window that can watch for console-minimize
events and hide the console window (minimize-to-tray).

- New option /hide_on_start.

- New option /verbose.

- Add a session instance mutex, so that only one instance of the program
  can be running at a time, per session.

- Add a synchronizing event (EventWaitHandle) to restore the console
  window if it is hidden.

This is so if a second instance is started in the same session then the
second can set the event to signal the first to restore its console
window, and then the second instance quits (as per session instance
mutex).

- Add a global mutex so that only one instance of the program can be
  syncing the brightness at a time.

This is so the program can be run by multiple users at the same time
without race conditions. If a WMI brightness event is received and
instance y is syncing the brightness then instance x waits to acquire
the mutex before checking/syncing.

- Remove Console.ReadKey to end the program when it is running normally.

Since the program must now enter a message loop to handle
synchronization events and window messages it no longer will exit when a
key is pressed in the console, under normal circumstances.

Closes https://github.com/HubKing/LinkBrightness/pull/2